requirements.txt
aiohttp[speedups]
aria2p
async_generator
asyncurban
bs4
CairoSVG
cffi
cowpy
dnspython
emoji
gitpython
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google_images_download
googletrans==4.0.0rc1
gTTS
gTTS-token
hachoir
heroku3
httplib2
deezloader
humanize
js2py
lottie
lxml
lyricsgenius
oauth2client
Pillow
psycopg2
psycopg2-binary
pybase64
pyfiglet
PyGithub
pylast
pySmartDL
python-barcode
python-dateutil
python-dotenv
pytz
qrcode
requests
search-engine-parser
speedtest-cli
sqlalchemy==1.3.23
telegraph
git+https://github.com/PhoenixError/Telethon
telethon-session-sqlalchemy
telethon-tgcrypto
wikipedia
youtube-dl
youtube_search
